Removed unnecessary header files from nav.c. Then I moved all of the #include statements from nav.c to nav.h, except for #include "nav.h". Verified it still compiles. Did the same thing for planner.
Also switched planner to use the timing library. This should really be double
checked to verify I didn't break anything. I will do it when I have time but
please be on the lookout for any funny business from planner.


git-svn-id: file:///home/thomas/tmp/stingray/libs@268 ae4af862-681a-43e1-b28c-8c040b23e139

+++ parser/include/parser.h
@@ -155,7 +155,7 @@ typedef struct _CONF_VARS {
     int         relay_baud;
     int         debug_level;
     int         op_mode;
-    int         enable_log;
+    float       enable_log;
 	float		depth_gate;
 	float		depth_pipe;
 	float		depth_buoy;
